+++ openghg_inversions/hbmcmc/inversion_pymc.py
@@ -18,7 +18,7 @@ from pytensor.tensor import TensorVariable
 
 from openghg_inversions import convert
 from openghg_inversions import utils
-from openghg_inversions.hbmcmc.inversionsetup import offset_matrix
+from openghg_inversions.hbmcmc.components import make_offset
 from openghg_inversions.hbmcmc.hbmcmc_output import define_output_filename
 from openghg_inversions.config.version import code_version
 
@@ -152,6 +152,8 @@ def inferpymc(
     reparameterise_log_normal: bool = False,
     pollution_events_from_obs: bool = False,
     no_model_error: bool = False,
+    offset_args: dict | None = None,
+    power: dict | float = 1.99,
 ) -> dict:
     """Uses PyMC module for Bayesian inference for emissions field, boundary
     conditions and (currently) a single model error value.
@@ -223,6 +225,10 @@ def inferpymc(
       no_model_error:
         When True, only use observation error in likelihood function (omitting min. model error
         and model error from scaling pollution events.)
+      offset_args: optional arguments to pass to `make_offset`.
+      power: power to raise pollution events to when using pollution events from obs. Default is 1.99.
+        Any value (strictly) between 1 and 2 will work. If a dictionary is passed, this is used to create
+        a prior for the power, making the power a hyper-parameter.
 
     Returns:
       Dictionary containing:
@@ -271,16 +277,7 @@ def inferpymc(
     nit = int(nit)
 
     # convert siteindicator into a site indexer
-    if sigma_per_site:
-        sites = siteindicator.astype(int)
-        nsites = np.amax(sites) + 1
-    else:
-        sites = np.zeros_like(siteindicator).astype(int)
-        nsites = 1
-    nsigmas = np.amax(sigma_freq_index) + 1
-
-    if add_offset:
-        B = offset_matrix(siteindicator)
+    sites = siteindicator.astype(int) if sigma_per_site else np.zeros_like(siteindicator).astype(int)
 
     coords = _make_coords(
         Y, Hx, siteindicator, sigma_freq_index, Hbc, sigma_per_site=sigma_per_site, sites=None
@@ -320,9 +317,8 @@ def inferpymc(
             mu += mu_bc
 
         if add_offset:
-            offset0 = parse_prior("offset0", offsetprior, shape=int(nsites - 1))
-            offset_vec = pt.concatenate((np.array([0]), offset0), axis=0)
-            offset = pm.Deterministic("offset", pt.dot(B, offset_vec), dims="nmeasure")
+            offset_args = offset_args or {}
+            offset = make_offset(siteindicator, offsetprior, **offset_args)
             mu += offset
 
         Y = pm.Data("Y", Y, dims="nmeasure")  # type: ignore
@@ -331,7 +327,7 @@ def inferpymc(
 
         if pollution_events_from_obs is True:
             if use_bc is True:
-                pollution_event = pt.abs(Y - pt.dot(hbc, bc))
+                pollution_event = pt.abs(Y - mu_bc)
             else:
                 pollution_event = pt.abs(Y) + 1e-6 * pt.mean(Y)  # small non-zero term to prevent NaNs
         else:
@@ -345,7 +341,8 @@ def inferpymc(
             small_amount = 1e-12 * mean_obs
             eps = pt.maximum(pt.abs(error), small_amount)  # type: ignore
         else:
-            eps = pt.maximum(pt.sqrt(error**2 + pollution_event_scaled_error**2), min_error)  # type: ignore
+            power0 = parse_prior("power", power) if isinstance(power, dict) else power
+            eps = pt.maximum(pt.sqrt(error**2 + pt.pow(pollution_event_scaled_error, power0)), min_error)  # type: ignore
 
         epsilon = pm.Deterministic("epsilon", eps, dims="nmeasure")
 
@@ -990,7 +987,7 @@ def inferpymc_postprocessouts(
             do_not_compress.append(dv)
 
     # setting compression levels for data vars in outds
-    comp = dict(zlib=True, complevel=5)
+    comp = dict(zlib=True, complevel=5, shuffle=True)
     encoding = {var: comp for var in outds.data_vars if var not in do_not_compress}
 
     output_filename = define_output_filename(outputpath, species, domain, outputname, start_date, ext=".nc")