+++ openghg_inversions/inversion_data/get_data.py
@@ -85,9 +85,9 @@ def add_obs_error(sites: list[str], fp_all: dict, add_averaging_error: bool = Tr
 
         elif add_averaging_error:
             # Fill with zeros so that if one of repeatability and variability is not NaN, then mf_error will not be NaN.
-            ds["mf_error"] = np.sqrt(ds["mf_repeatability"].fillna(0) ** 2 + ds["mf_variability"].fillna(0) ** 2)
-            # Fill "mf_error" with nans if repeatability and variability are both NaN
-            ds["mf_error"] = ds["mf_error"].where(~(np.isnan(ds["mf_repeatability"]) & np.isnan(ds["mf_variability"])))
+            ds["mf_error"] = np.sqrt(
+                ds["mf_repeatability"].fillna(0) ** 2 + ds["mf_variability"].fillna(0) ** 2
+            )
         else:
             ds["mf_error"] = ds["mf_repeatability"]
 
@@ -95,11 +95,26 @@ def add_obs_error(sites: list[str], fp_all: dict, add_averaging_error: bool = Tr
         ds["mf_error"].attrs["units"] = ds.mf.attrs.get("units", None)
 
         # warnings/info for debugging
-        err0 = ds["mf_error"] == 0
+        err0 = (ds["mf_error"] == 0) | (
+            ds["mf_error"].isnull()
+        )  # might have NaN if add_averaging_error is False
 
         if err0.any():
             percent0 = 100 * err0.mean()
-            logger.warning("`mf_error` is zero for %.0f percent of times at site %s.", percent0, site)
+            logger.warning("`mf_error` is zero/nan for %.2f percent of times at site %s.", percent0, site)
+
+            if percent0 > 10:
+                mf_err_da = ds["mf_error"].as_numpy()  # load into memory to avoid Dask issues
+                fill_value = np.nanmax(
+                    [
+                        mf_err_da.where(mf_err_da != 0).dropna(dim="time").median(),
+                        ds["mf"].std(dim="time"),
+                    ]
+                )
+                ds["mf_error"] = mf_err_da.where(mf_err_da != 0, fill_value)
+                logger.warning(
+                    "More than 10% of `mf_error` is zero/nan, it is thus filled with the max of median `mf_error` and stdev of `mf`."
+                )
             info_msg = (
                 "If `averaging_period` matches the frequency of the obs data, then `mf_variability` "
                 "will be zero. Try setting `averaging_period = None`."
@@ -107,7 +122,9 @@ def add_obs_error(sites: list[str], fp_all: dict, add_averaging_error: bool = Tr
             logger.info(info_msg)
 
 
-def convert_to_list(x: list[str | None] | str | None, length: int, name: str | None = None) -> list[str | None]:
+def convert_to_list(
+    x: list[str | None] | str | None, length: int, name: str | None = None
+) -> list[str | None]:
     """Convert variable that might be list, str, or None to a list of the expected size.
 
     Args:
@@ -307,13 +324,14 @@ def data_processing_surface_notracer(
     units = {}
     site_indices_to_keep = []
 
-    keep_variables = [f"{species}",
-                        f"{species}_variability", 
-                        f"{species}_repeatability",
-                        f"{species}_number_of_observations",
-                        "inlet",  # needed if multiple inlets combined
-                        "inlet_height",  # sometimes needed if inlet='multiple' (may be outdated soon)
-                        ]
+    keep_variables = [
+        f"{species}",
+        f"{species}_variability",
+        f"{species}_repeatability",
+        f"{species}_number_of_observations",
+        "inlet",  # needed if multiple inlets combined
+        "inlet_height",  # sometimes needed if inlet='multiple' (may be outdated soon)
+    ]
     warnings.warn(f"Dropping all variables besides {keep_variables}")
     for i, site in enumerate(sites):
         # Get observations data
@@ -328,7 +346,7 @@ def data_processing_surface_notracer(
             instrument=instrument[i],
             calibration_scale=calibration_scale,
             stores=obs_store,
-            keep_variables=keep_variables
+            keep_variables=keep_variables,
         )
 
         if site_data is None: