+++ openghg_inversions/postprocessing/countries.py
@@ -309,7 +309,7 @@ class Countries:
         """
         country_file_path = get_country_file_path(country_file=country_file, domain=domain)
         return cls(
-            xr.open_dataset(country_file_path),
+            xr.open_dataset(country_file_path, engine="h5netcdf"),
             country_code=country_code,
             country_selections=country_selections,
             country_regions=country_regions,